requirements.txt
openai
ffmpeg-python